+++ src/main.cpp
@@ -25,7 +25,7 @@ int main(int argc, char *argv[])
 
   App.setOrganizationName("ChungZH");
   App.setApplicationName("Notepanda");
-  App.setApplicationVersion("0.0.2");
+  App.setApplicationVersion("0.0.3");
 
   QCommandLineParser parser;
   parser.addHelpOption();