+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Custom syntax highlight theme (light / dark)
 - Set window modified
+- New logo
 
 ### 🐞 Bug fix