+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
 - Set window modified
 - New logo
 - More available syntax highlight themes: `Breeze Dark`, `Default`, `Printing`, `Solarized Dark` and `Solarized  Light`
+- Sticky mode
+    - Hide sidebar
+    - Hide line number area
 
 ### 🐞 Bug fix